Rework dir and commit boxes.
Introduce a ChunkContainerRef and add config information for chunk
container information into the ref.

+++ core/src/main/java/org/fejoa/chunkstore/IChunkPointer.java
@@ -14,7 +14,7 @@ import java.io.IOException;
 
 public interface IChunkPointer {
     int getPointerLength();
-    int getDataLength();
+    long getDataLength();
     BoxPointer getBoxPointer();
     void setBoxPointer(BoxPointer boxPointer);
     void read(DataInputStream inputStream) throws IOException;
